branches/develop-node/marmot/requirements.txt
Django==1.8.17
django-guardian==1.4.6
django-redis==4.7.0
celery==3.1.25
MySQL-python==1.2.5
redis==2.10.5
redis-py-cluster==1.2.0
requests==2.13.0
zeroc-ice==3.6.3